backend/requirements.txt
fastapi
uvicorn
pydantic-ai[logfire]
httpx
python-dotenv
